DOP-2123: Rename Alert to Banner
